modify directory
